Decent chuck of updates to nav
