+++ lib/DelayData.tsx
@@ -1,0 +1,3 @@
+export const delayData = <T, >(value: T): Promise<T> => {
+  return new Promise((resolve) => setTimeout(() => resolve(value), 1000))
+}
